can upload study rules
